+++ unstructured_inference/inference/layoutelement.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from dataclasses import dataclass
-from typing import Collection, List, Optional
+from typing import Any, Collection, Iterable, List, Optional
 
 import numpy as np
 from layoutparser.elements.layout import TextBlock
@@ -19,10 +19,85 @@ from unstructured_inference.inference.elements import (
     ImageTextRegion,
     Rectangle,
     TextRegion,
+    TextRegions,
+    coords_intersections,
     grow_region_to_match_region,
-    intersections,
     region_bounding_boxes_are_almost_the_same,
 )
+
+EPSILON_AREA = 1e-7
+
+
+@dataclass
+class LayoutElements(TextRegions):
+    element_probs: np.ndarray = np.array([])
+    element_class_ids: np.ndarray = np.array([])
+    element_class_id_map: dict[int, str] | None = None
+
+    def __post_init__(self):
+        if self.element_probs is not None:
+            self.element_probs = self.element_probs.astype(float)
+        element_size = self.element_coords.shape[0]
+        for attr in ("element_probs", "element_class_ids", "texts"):
+            if getattr(self, attr).size == 0 and element_size:
+                setattr(self, attr, np.array([None] * element_size))
+
+    def slice(self, indices) -> LayoutElements:
+        """slice and return only selected indices"""
+        return LayoutElements(
+            element_coords=self.element_coords[indices],
+            texts=self.texts[indices],
+            source=self.source,
+            element_probs=self.element_probs[indices],
+            element_class_ids=self.element_class_ids[indices],
+            element_class_id_map=self.element_class_id_map,
+        )
+
+    @classmethod
+    def concatenate(cls, groups: Iterable[LayoutElements]) -> LayoutElements:
+        """concatenate a sequence of LayoutElements in order as one LayoutElements"""
+        coords, texts, probs, class_ids = [], [], [], []
+        class_id_map = {}
+        for group in groups:
+            coords.append(group.element_coords)
+            texts.append(group.texts)
+            probs.append(group.element_probs)
+            class_ids.append(group.element_class_ids)
+            if group.element_class_id_map:
+                class_id_map.update(group.element_class_id_map)
+        return cls(
+            element_coords=np.concatenate(coords),
+            texts=np.concatenate(texts),
+            element_probs=np.concatenate(probs),
+            element_class_ids=np.concatenate(class_ids),
+            element_class_id_map=class_id_map,
+            source=group.source,
+        )
+
+    def as_list(self):
+        """return a list of LayoutElement for backward compatibility"""
+        return [
+            LayoutElement.from_coords(
+                x1,
+                y1,
+                x2,
+                y2,
+                text=text,
+                type=(
+                    self.element_class_id_map[class_id]
+                    if class_id and self.element_class_id_map
+                    else None
+                ),
+                prob=prob,
+                source=self.source,
+            )
+            for (x1, y1, x2, y2), text, prob, class_id in zip(
+                self.element_coords,
+                self.texts,
+                self.element_probs,
+                self.element_class_ids,
+            )
+        ]
 
 
 @dataclass
@@ -215,7 +290,10 @@ def separate(region_a: Rectangle, region_b: Rectangle):
 
 
 def table_cells_to_dataframe(
-    cells: List[dict], nrows: int = 1, ncols: int = 1, header=None
+    cells: List[dict],
+    nrows: int = 1,
+    ncols: int = 1,
+    header=None,
 ) -> DataFrame:
     """convert table-transformer's cells data into a pandas dataframe"""
     arr = np.empty((nrows, ncols), dtype=object)
@@ -232,23 +310,172 @@ def table_cells_to_dataframe(
     return DataFrame(arr, columns=header)
 
 
-def partition_groups_from_regions(regions: Collection[TextRegion]) -> List[List[TextRegion]]:
+def partition_groups_from_regions(regions: TextRegions) -> List[TextRegions]:
     """Partitions regions into groups of regions based on proximity. Returns list of lists of
     regions, each list corresponding with a group"""
     if len(regions) == 0:
         return []
-    padded_regions = [
-        r.bbox.vpad(r.bbox.height * inference_config.ELEMENTS_V_PADDING_COEF).hpad(
-            r.bbox.height * inference_config.ELEMENTS_H_PADDING_COEF,
-        )
-        for r in regions
-    ]
+    padded_coords = regions.element_coords.copy()
+    v_pad = (regions.y2 - regions.y1) * inference_config.ELEMENTS_V_PADDING_COEF
+    h_pad = (regions.x2 - regions.x1) * inference_config.ELEMENTS_H_PADDING_COEF
+    padded_coords[:, 0] -= h_pad
+    padded_coords[:, 1] -= v_pad
+    padded_coords[:, 2] += h_pad
+    padded_coords[:, 3] += v_pad
 
-    intersection_mtx = intersections(*padded_regions)
+    intersection_mtx = coords_intersections(padded_coords)
 
-    _, group_nums = connected_components(intersection_mtx)
-    groups: List[List[TextRegion]] = [[] for _ in range(max(group_nums) + 1)]
-    for region, group_num in zip(regions, group_nums):
-        groups[group_num].append(region)
+    group_count, group_nums = connected_components(intersection_mtx)
+    groups: List[TextRegions] = []
+    for group in range(group_count):
+        groups.append(regions.slice(np.where(group_nums == group)[0]))
 
     return groups
+
+
+def intersection_areas_between_coords(
+    coords1: np.ndarray,
+    coords2: np.ndarray,
+    threshold: float = 0.5,
+):
+    """compute intersection area and own areas for two groups of bounding boxes"""
+    x11, y11, x12, y12 = np.split(coords1, 4, axis=1)
+    x21, y21, x22, y22 = np.split(coords2, 4, axis=1)
+
+    xa = np.maximum(x11, np.transpose(x21))
+    ya = np.maximum(y11, np.transpose(y21))
+    xb = np.minimum(x12, np.transpose(x22))
+    yb = np.minimum(y12, np.transpose(y22))
+
+    return np.maximum((xb - xa), 0) * np.maximum((yb - ya), 0)
+
+
+def clean_layoutelements(elements: LayoutElements, subregion_threshold: float = 0.5):
+    """After this function, the list of elements will not contain any element inside
+    of the type specified"""
+    # Sort elements from biggest to smallest
+    if len(elements) < 2:
+        return elements
+
+    sorted_by_area = np.argsort(-elements.areas)
+    sorted_coords = elements.element_coords[sorted_by_area]
+
+    # First check if targets contains each other
+    self_intersection = intersection_areas_between_coords(sorted_coords, sorted_coords)
+    areas = elements.areas[sorted_by_area]
+    # check from largest to smallest regions to find if it contains any other regions
+    is_almost_subregion_of = (
+        self_intersection / np.maximum(areas, EPSILON_AREA) > subregion_threshold
+    ) & (areas <= areas.T)
+
+    n_candidates = len(elements)
+    mask = np.ones_like(areas, dtype=bool)
+    current_candidate = 0
+    while n_candidates > 1:
+        plus_one = current_candidate + 1
+        remove = (
+            np.where(is_almost_subregion_of[current_candidate, plus_one:])[0]
+            + current_candidate
+            + 1
+        )
+
+        if not remove.sum():
+            break
+
+        mask[remove] = 0
+        n_candidates -= len(remove) + 1
+        remaining_candidates = np.where(mask[plus_one:])[0]
+
+        if not len(remaining_candidates):
+            break
+
+        current_candidate = remaining_candidates[0] + plus_one
+
+    final_coords = sorted_coords[mask]
+    sorted_by_y1 = np.argsort(final_coords[:, 1])
+
+    final_attrs: dict[str, Any] = {"element_class_id_map": elements.element_class_id_map}
+    for attr in ("element_class_ids", "element_probs", "texts"):
+        if (original_attr := getattr(elements, attr)) is None:
+            continue
+        final_attrs[attr] = original_attr[sorted_by_area][mask][sorted_by_y1]
+    final_elements = LayoutElements(element_coords=final_coords[sorted_by_y1], **final_attrs)
+    return final_elements
+
+
+def clean_layoutelements_for_class(
+    elements: LayoutElements,
+    element_class: int,
+    subregion_threshold: float = 0.5,
+):
+    """After this function, the list of elements will not contain any element inside
+    of the type specified"""
+    # Sort elements from biggest to smallest
+    sorted_by_area = np.argsort(-elements.areas)
+    sorted_coords = elements.element_coords[sorted_by_area]
+
+    target_indices = elements.element_class_ids[sorted_by_area] == element_class
+
+    # skip trivial result
+    len_target = target_indices.sum()
+    if len_target == 0 or len_target == len(elements):
+        return elements
+
+    target_coords = sorted_coords[target_indices]
+    other_coords = sorted_coords[~target_indices]
+
+    # First check if targets contains each other
+    target_self_intersection = intersection_areas_between_coords(target_coords, target_coords)
+    target_areas = elements.areas[sorted_by_area][target_indices]
+    # check from largest to smallest regions to find if it contains any other regions
+    is_almost_subregion_of = (
+        target_self_intersection / np.maximum(target_areas, EPSILON_AREA) > subregion_threshold
+    ) & (target_areas <= target_areas.T)
+
+    n_candidates = len_target
+    mask = np.ones_like(target_areas, dtype=bool)
+    current_candidate = 0
+    while n_candidates > 1:
+        plus_one = current_candidate + 1
+        remove = (
+            np.where(is_almost_subregion_of[current_candidate, plus_one:])[0]
+            + current_candidate
+            + 1
+        )
+
+        if not remove.sum():
+            break
+
+        mask[remove] = 0
+        n_candidates -= len(remove) + 1
+        remaining_candidates = np.where(mask[plus_one:])[0]
+
+        if not len(remaining_candidates):
+            break
+
+        current_candidate = remaining_candidates[0] + plus_one
+
+    target_coords_to_keep = target_coords[mask]
+
+    other_to_target_intersection = intersection_areas_between_coords(
+        other_coords,
+        target_coords_to_keep,
+    )
+    # check from largest to smallest regions to find if it contains any other regions
+    other_areas = elements.areas[sorted_by_area][~target_indices]
+    other_is_almost_subregion_of_target = (
+        other_to_target_intersection / np.maximum(other_areas, EPSILON_AREA) > subregion_threshold
+    ) & (other_areas.reshape((-1, 1)) <= target_areas[mask].T)
+
+    other_mask = ~other_is_almost_subregion_of_target.sum(axis=1).astype(bool)
+
+    final_coords = np.vstack([target_coords[mask], other_coords[other_mask]])
+    final_attrs: dict[str, Any] = {"element_class_id_map": elements.element_class_id_map}
+    for attr in ("element_class_ids", "element_probs", "texts"):
+        if (original_attr := getattr(elements, attr)) is None:
+            continue
+        final_attrs[attr] = np.concatenate(
+            (original_attr[target_indices][mask], original_attr[~target_indices][other_mask]),
+        )
+    final_elements = LayoutElements(element_coords=final_coords, **final_attrs)
+    return final_elements